requirements.txt
numpy>=1.24,<3
scipy>=1.10
matplotlib>=3.7
litellm>=1.50
python-dotenv>=1.0
